# book-network/src/main/java/com/davidnhn/book/feedback/FeedbackResponse.java
package com.davidnhn.book.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private  Double note;
    private String comment;
    private boolean ownFeedback;
}